Fix password reset URLs when using a custom model
When using a custom `user_model` other than `User`, generating the
password reset link and attempting to reset the password would raise an
error. This is because the URL helpers were hardcoded to use the `User`
model, such as `edit_user_password_url`.

This commit uses Rails' [PolymorphicRoutes] to build the URL from the
passed in model. This will use the record's model to generate the URL,
allowing the URL to respect any custom model.

[PolymorphicRoutes]: https://api.rubyonrails.org/classes/ActionDispatch/Routing/PolymorphicRoutes.html